+++ src/extension.ts
@@ -2,9 +2,9 @@ import * as vscode from 'vscode';
 import * as fs from "fs";
 import * as path from "path";
 import * as mime from "mime";
-import { convertBytes, dirSize, formatDate, imageDimension, toString, musicMetaData, videoMetaData } from './helpers';
 import { Settings } from './Settings';
 import * as humanizeDuration from "humanize-duration";
+import { convertBytes, formatDate, getAudioDetails, getImageDetails, getStats, getVideoDetails, toString } from './helpers';
 
 
 export function activate(context: vscode.ExtensionContext) {
@@ -17,72 +17,68 @@ export function activate(context: vscode.ExtensionContext) {
 			location: vscode.ProgressLocation.Notification,
 			title: "Please wait...",
 		}, async () => {
-			const stat = fs.statSync(fsPath);
-			const extension = path.extname(fsPath) || "[none]";
-			const title = path.basename(fsPath);
-			const name = path.basename(fsPath, extension);
-			const type = stat.isFile() ? "File" : "Folder";
-			const directory = path.dirname(fsPath);
-			const size = stat.isFile() ? convertBytes(stat.size) || 0 : convertBytes(await dirSize(fsPath));
-			const created = formatDate(stat.birthtime) || "";
-			const changed = formatDate(stat.ctime) || "";
-			const modified = formatDate(stat.mtime) || "";
-			const accessed = formatDate(stat.atime) || "";
-			const mimeType = mime.getType(fsPath) || '[unknown]';
-			const dimensions = mimeType.includes('image') ? imageDimension(fsPath) : "";
-			const musicmetadata = mimeType.includes('audio') ? await musicMetaData(fsPath) : "";
-			const videometadata = mimeType.includes('video') ? await videoMetaData(fsPath) : "";
 
-			const baseDetails = toString([
-				["Name", type === 'File' ? name : title],
-				["Extension", extension, type === 'File'],
-				["Type", type],
-				["Size", size]
+			const stats = await getStats(fsPath);
+
+			const mimeType = mime.getType(fsPath) || '[unknown]';
+
+			const image = mimeType.includes('image') ? getImageDetails(fsPath) : "";
+
+			const audio = mimeType.includes('audio') ? await getAudioDetails(fsPath) : "";
+
+			const video = mimeType.includes('video') ? await getVideoDetails(fsPath) : "";
+
+			const pathDetails = toString([
+				["Name", stats.isFile ? stats.baseName : stats.fileName],
+				["Extension", stats.extension, stats.isFile],
+				["Type", stats.type],
+				["Size", convertBytes(stats.size), stats.size],
+				["Contains", `${stats.contains?.files} Files, ${stats.contains?.folders} Folders`, !stats.isFile && stats.contains]
 			]);
 
-			const dimensionDetails = dimensions ?
+			const imageDetails = image ?
 				toString([
-					["Dimensions", `${dimensions?.width} x ${dimensions?.height} pixels`],
-					["Width", `${dimensions?.width} pixels`],
-					["Height", `${dimensions?.height} pixels`],
+					["Dimensions", `${image?.width} x ${image?.height} pixels`],
+					["Width", `${image?.width} pixels`],
+					["Height", `${image?.height} pixels`],
 				]) : "";
 
-			const audioDetails = musicmetadata ?
+			const audioDetails = audio ?
 				toString([
-					["Title", musicmetadata.title, Settings.audioMetaData.title],
-					["Album", musicmetadata.album, Settings.audioMetaData.album],
-					["Artist", musicmetadata.artist.join(', '), Settings.audioMetaData.artist],
-					["Genre", musicmetadata.genre.join(', '), Settings.audioMetaData.genre],
-					["Year", musicmetadata.year, Settings.audioMetaData.year],
-					["Duration", humanizeDuration(musicmetadata.duration * 1000, { maxDecimalPoints: 2 }), Settings.showDuration && musicmetadata.duration],
+					["Title", audio.title, Settings.audioMetaData.title],
+					["Album", audio.album, Settings.audioMetaData.album],
+					["Artist", audio.artist.join(', '), Settings.audioMetaData.artist],
+					["Genre", audio.genre.join(', '), Settings.audioMetaData.genre],
+					["Year", audio.year, Settings.audioMetaData.year],
+					["Duration", humanizeDuration(audio.duration * 1000, { maxDecimalPoints: 2 }), Settings.showDuration && audio.duration],
 				]) : "";
 
-			const videoDetails = videometadata ?
+			const videoDetails = video ?
 				toString([
-					["Duration", humanizeDuration(videometadata.duration * 1000, { maxDecimalPoints: 2 }), Settings.showDuration && videometadata.duration],
+					["Duration", humanizeDuration(video.duration * 1000, { maxDecimalPoints: 2 }), Settings.showDuration && video.duration],
 				]) : "";
 
 			const locationDetails = toString([
-				['Location', args.fsPath],
-				['Directory', directory],
+				['Directory', stats.directory],
+				['Location', stats.location],
 			]);
 
 			const timestampDetails = toString([
-				['Created', created, Settings.timeStamps.createdTimestamp],
-				['Changed', changed, Settings.timeStamps.changedTimestamp],
-				['Modified', modified, Settings.timeStamps.modifiedTimestamp],
-				['Accessed', accessed, Settings.timeStamps.accessedTimestamp]
+				['Created', formatDate(stats.created), stats.created && Settings.timeStamps.createdTimestamp],
+				['Changed', formatDate(stats.changed), stats.changed && Settings.timeStamps.changedTimestamp],
+				['Modified', formatDate(stats.modified), stats.modified && Settings.timeStamps.modifiedTimestamp],
+				['Accessed', formatDate(stats.accessed), stats.accessed && Settings.timeStamps.accessedTimestamp]
 			]);
 
 			const mimeTypeDetails = toString([
-				['Mime Type', mimeType, type === 'File']
+				['Mime Type', mimeType, stats.isFile]
 			]);
 
-			const result = [baseDetails, dimensionDetails, audioDetails, videoDetails, locationDetails, timestampDetails, mimeTypeDetails]
+			const result = [pathDetails, imageDetails, audioDetails, videoDetails, locationDetails, timestampDetails, mimeTypeDetails]
 				.filter(Boolean)
 				.join('\n----------------------------------------------------------------------\n');
 
-			vscode.window.showInformationMessage(title, { modal: true, detail: result });
+			vscode.window.showInformationMessage(stats.fileName, { modal: true, detail: result });
 		});
 
 	});