+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [v5.0.0]
+
+- Now it shows number of folders and files inside the folder properties.
+
+## [v4.0.0]
+
+- Added Audio and Video Details.
+
 ## [v3.0.0]
 
 - Hide extension and mimetype for folder properties.